+++ beancount2ledger/__init__.py
@@ -4,9 +4,9 @@ Beancount to Ledger converter
 
 # SPDX-FileCopyrightText: © 2020 Software in the Public Interest, Inc.
 
-# SPDX-License-Identifier: GPL-2.0-only
+# SPDX-License-Identifier: GPL-2.0-or-later
 
-__license__ = "GPL-2.0-only"
+__license__ = "GPL-2.0-or-later"
 
 from beancount import loader