+++ beancount2ledger/cli.py
@@ -5,9 +5,9 @@ Beancount to Ledger converter
 # SPDX-FileCopyrightText: © 2020 Software in the Public Interest, Inc.
 # SPDX-FileCopyrightText: © 2020 Martin Michlmayr <tbm@cyrius.com>
 
-# SPDX-License-Identifier: GPL-2.0-only
+# SPDX-License-Identifier: GPL-2.0-or-later
 
-__license__ = "GPL-2.0-only"
+__license__ = "GPL-2.0-or-later"
 
 import argparse
 import sys