+++ beancount2ledger/common.py
@@ -4,9 +4,9 @@ Common functions
 
 # SPDX-FileCopyrightText: © 2014-2017 Martin Blais
 
-# SPDX-License-Identifier: GPL-2.0-only
+# SPDX-License-Identifier: GPL-2.0-or-later
 
-__license__ = "GPL-2.0-only"
+__license__ = "GPL-2.0-or-later"
 
 import re