+++ beancount2ledger/hledger.py
@@ -4,9 +4,9 @@ Convert Beancount entries to hledger
 
 # SPDX-FileCopyrightText: © 2014-2017 Martin Blais
 
-# SPDX-License-Identifier: GPL-2.0-only
+# SPDX-License-Identifier: GPL-2.0-or-later
 
-__license__ = "GPL-2.0-only"
+__license__ = "GPL-2.0-or-later"
 
 from beancount.core.amount import Amount
 from beancount.core import position